Add getter contract functions to get committed and owned pixels

+++ contracts/src/systems/guilds.cairo
@@ -8,8 +8,8 @@ pub trait IGuild<T> {
     fn remove_member(ref self: T, game_id: u32, guild_id: u32, member: ContractAddress);
     fn is_member(ref self: T, game_id: u32, guild_id: u32, member: ContractAddress) -> bool;
     fn get_guild_contract_address(ref self: T) -> ContractAddress;
-    fn get_player_commit(ref self: T, player_address: ContractAddress) -> u16;
-    fn get_player_owns(ref self: T, player_address: ContractAddress) -> u16;
+    fn get_player_commit(ref self: T, player_address: ContractAddress) -> u32;
+    fn get_player_owns(ref self: T, player_address: ContractAddress) -> u32;
     fn get_guild_points(ref self: T, game_id: u32, guild_id: u32) -> u32;
 }
 
@@ -223,14 +223,14 @@ mod guild_actions {
             guild_contract_address
         }
 
-        fn get_player_commit(ref self: ContractState, player_address: ContractAddress) -> u16 {
+        fn get_player_commit(ref self: ContractState, player_address: ContractAddress) -> u32 {
             // let mut core_world = self.world(@"pixelaw");
             let mut world = self.world(@"p_war");
             let mut player: Player = world.read_model(player_address);
             player.num_commit
         }
 
-        fn get_player_owns(ref self: ContractState, player_address: ContractAddress) -> u16 {
+        fn get_player_owns(ref self: ContractState, player_address: ContractAddress) -> u32 {
             // let mut core_world = self.world(@"pixelaw");
             let mut world = self.world(@"p_war");
             let mut player: Player = world.read_model(player_address);
